Switch to uniformly-sized blocks
Instead of changing `tasks_per_node` and `nodes_per_block` on each
iteration, use a uniformly-sized block and add a block on each
iteration. The advantage of this is that it allows for us to separately
measure overhead due to the number of workers running on a single node,
and the overhead due to running many total workers. Other minor changes:
remove unused variables and imports, print exceptions, and write out the
number of requested blocks.